Fetching data

+++ techbootcamp-zach/src/main/java/com/perficient/techbootcampzach/map/MapArea.java
@@ -1,7 +1,7 @@
 package com.perficient.techbootcampzach.map;
 
 public interface MapArea {
-	public boolean contains(int x, int y);
+	public boolean contains(double x, double y);
 	public String getCallout();
 	public void setCallout(String callout);
 }
